+++ wagtail_localize/modeladmin/views.py
@@ -1,4 +1,3 @@
-from django import VERSION as DJANGO_VERSION
 from django.apps import apps
 from django.conf import settings
 from django.contrib.admin.utils import unquote
@@ -65,12 +64,9 @@ class TranslatableIndexView(TranslatableViewMixin, IndexView):
     def get_filters(self, request):
         filters = super().get_filters(request)
         # Update the 'lookup_params' part of the filters tuple to filter objects
-        # using the currently active locale
-        if DJANGO_VERSION >= (5, 0):
-            # modeladmin IndexView needs an iterable in build_q_object_from_lookup_parameters
-            filters[2]["locale"] = [self.locale.id]
-        else:
-            filters[2]["locale"] = self.locale.id
+        # using the currently active locale.
+        # modeladmin IndexView needs an iterable in build_q_object_from_lookup_parameters
+        filters[2]["locale"] = [self.locale.id]
         return filters
 
     def get_context_data(self, **kwargs):