+++ wagtail_localize/segments/tests/test_segment_ingestion.py
@@ -2,7 +2,6 @@ import unittest
 import uuid
 
 from django.test import TestCase
-from wagtail import VERSION as WAGTAIL_VERSION
 from wagtail.blocks import StreamValue
 from wagtail.images import get_image_model
 from wagtail.images.tests.utils import get_test_image_file
@@ -560,9 +559,6 @@ class TestSegmentIngestionWithStreamField(TestCase):
             ],
         )
 
-    @unittest.skipUnless(
-        WAGTAIL_VERSION >= (6, 3), "ImageBlock was added in Wagtail 6.3"
-    )
     def test_imageblock(self):
         block_id = uuid.uuid4()
         test_image = get_image_model().objects.create(